# chronospace-be/docs/swagger.yaml
basePath: /
definitions:
...
  chronospace-be_internal_models.Booking:
    properties:
      date:
        $ref: '#/definitions/pgtype.Date'
      id:
        type: string
      service_id:
        type: string
      status:
        type: string
      time:
        type: string
      user_id:
        type: string
    type: object
  chronospace-be_internal_models.CreateBookingParams:
    properties:
      date:
        $ref: '#/definitions/pgtype.Date'
      service_id:
        type: string
      status:
        type: string
      time:
        type: string
      user_id:
        type: string
    type: object
  chronospace-be_internal_models.CreateScheduleRequest:
    properties:
      end_time:
        type: string
      start_time:
        type: string
    required:
    - end_time
    - start_time
    type: object
...
  chronospace-be_internal_models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  chronospace-be_internal_models.ScheduleResponse:
    properties:
      id:
        type: string
      status:
        type: string
      time_end:
        type: string
      time_start:
        type: string
    type: object
...
  chronospace-be_internal_models.UpdateBookingParams:
    properties:
      date:
        $ref: '#/definitions/pgtype.Date'
      id:
        type: string
      status:
        type: string
      time:
        type: string
    type: object
  chronospace-be_internal_models.UpdateScheduleRequest:
    properties:
      end_time:
        type: string
      start_time:
        type: string
    type: object
...
  pgtype.Date:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
host: localhost:8080
info:
  contact:
    email: support@chronospace.com
    name: API Support
    url: http://www.chronospace.com/support
  description: This is the backend server for the Chronospace application.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Chronospace API
  version: "1.0"
paths:
  /schedules/{id}:
    delete:
      description: Delete a schedule by its ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Delete a schedule
      tags:
      - schedules
  /v1/api/bookings:
    get:
      consumes:
      - application/json
      description: Get all bookings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chronospace-be_internal_models.Booking'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: List all bookings
      tags:
      - Booking
    post:
      consumes:
      - application/json
      description: Create a new booking
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/chronospace-be_internal_models.CreateBookingParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Create booking
      tags:
      - Booking
  /v1/api/bookings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Delete booking
      tags:
      - Booking
    get:
      consumes:
      - application/json
      description: Get booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Get booking
      tags:
      - Booking
    put:
      consumes:
      - application/json
      description: Update an existing booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Booking details
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/chronospace-be_internal_models.UpdateBookingParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.Booking'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Update booking
      tags:
      - Booking
  /v1/api/bookings/user:
    get:
      consumes:
      - application/json
      description: Get all bookings for the authenticated user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chronospace-be_internal_models.Booking'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: List user bookings
      tags:
      - Booking
  /v1/api/maps/search:
    get:
      consumes:
      - application/json
      description: Search for places using Google Maps API
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Search places
      tags:
      - Maps
  /v1/api/schedules:
    get:
      description: Get a list of all schedules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/chronospace-be_internal_models.ScheduleResponse'
            type: array
      summary: List all schedules
      tags:
      - schedules
    post:
      consumes:
      - application/json
      description: Create a new schedule with the given time range
      parameters:
      - description: Schedule details
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/chronospace-be_internal_models.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Create a new schedule
      tags:
      - schedules
  /v1/api/schedules/{id}:
    get:
      description: Get schedule details by its ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ScheduleResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Get a schedule by ID
      tags:
      - schedules
    put:
      consumes:
      - application/json
      description: Update a schedule's details by its ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Schedule details
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/chronospace-be_internal_models.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/chronospace-be_internal_models.ErrorResponse'
      summary: Update a schedule
      tags:
      - schedules
...
swagger: "2.0"